+++ packages/myel-client/src/__tests__/fixtures.ts
@@ -1,4 +1,6 @@
 import {bytes} from 'multiformats';
+import fs from 'fs';
+import BufferList from 'bl/BufferList';
 
 export const gsMsg1 = bytes.fromHex(
   '1aaa04100e1ab4010a2666696c2f646174612d7472616e736665722f696e636f6d696e672d726571756573742f312e31128901a36449735271f46752657175657374f668526573706f6e7365a66454797065006441637074f56450617573f4665866657249441b0000017b0bb0870d6456526573a466537461747573066249441b0000017b0bb0870d6b5061796d656e744f77656440674d6573736167656064565479707752657472696576616c4465616c526573706f6e73652f311aa3010a1566696c2f646174612d7472616e736665722f312e31128901a36449735271f46752657175657374f668526573706f6e7365a66454797065006441637074f56450617573f4665866657249441b0000017b0bb0870d6456526573a466537461747573066249441b0000017b0bb0870d6b5061796d656e744f77656440674d6573736167656064565479707752657472696576616c4465616c526573706f6e73652f311a680a1166696c2f646174612d7472616e73666572125383f4f68600f5f41b0000017b0bb0870da466537461747573066249441b0000017b0bb0870d6b5061796d656e744f77656440674d657373616765607752657472696576616c4465616c526573706f6e73652f311a5f0a1b677261706873796e632f726573706f6e73652d6d65746164617461124081a2646c696e6bd82a5827000171a0e402200a2439495cfb5eafbb79669f644ca2c5a3d31b28e96c424cde5dd0e540a7d9486c626c6f636b50726573656e74f522610a060171a0e402201257a16b446c5f69636f6e2e737667a3634b65796b446c5f69636f6e2e7376676556616c7565d82a5827000155a0e40220ff51b469c4722121632236eee2a7c20325a5c03c4a0b3739958a02e5b701ccb56453697a65190467'
@@ -23,3 +25,15 @@ export const gsMsgSingleBlock = bytes.fromHex(
 export const dtMsgSingleBlockComplete = bytes.fromHex(
   `a36449735271f46752657175657374f668526573706f6e7365a66454797065036441637074f56450617573f4665866657249441b0000017b9e995cd06456526573a4665374617475730f6249441b0000017b9e995cd06b5061796d656e744f77656440674d6573736167656064565479707752657472696576616c4465616c526573706f6e73652f31`
 );
+
+export async function* blockSource(
+  start: number,
+  end: number
+): AsyncIterable<BufferList> {
+  for (let i = start; i < end + 1; i++) {
+    const data = fs.readFileSync('src/__tests__/fixtures/chunk' + i + '.txt');
+    const bl = new BufferList();
+    bl.append(Buffer.from(data.toString(), 'hex'));
+    yield bl;
+  }
+}
