+++ packages/myel-client/src/selectors.ts
@@ -5,9 +5,6 @@ import * as dagCBOR from '@ipld/dag-cbor';
 import {sha256} from 'multiformats/hashes/sha2';
 import {Blockstore} from 'interface-blockstore';
 import {equals} from './filaddress';
-import {UnixFS} from 'ipfs-unixfs';
-import * as dagJSON from 'multiformats/codecs/json';
-import {PBLink} from '@ipld/dag-pb';
 
 enum Kind {
   Invalid = '',
@@ -723,36 +720,8 @@ export function decoderFor(cid: CID): Decoder | null {
   }
 }
 
-export function toPathComponents(path = ''): string[] {
-  // split on / unless escaped with \
-  return (path.trim().match(/([^\\^/]|\\\/)+/g) || []).filter(Boolean);
-}
-
-interface DAGResolver {
-  loadOrRequest(
-    root: CID,
-    link: CID,
-    sel: SelectorNode,
-    blk: CID
-  ): Promise<Block<any>>;
-}
-
-export async function* traverse(
-  root: CID, // the root node of the FULL DAG
-  link: CID, // the link at where we start the traversal. Can be the root.
-  sel: SelectorNode,
-  resolver: DAGResolver
-) {
-  const loader = {
-    load: (cid: CID): Promise<Block<any>> => {
-      return resolver.loadOrRequest(root, link, sel, cid);
-    },
-    close: () => {},
-  };
-  yield* walk(new Node(link), parseContext().parseSelector(sel), loader);
-}
-
-export async function* walk(
+// walkBlocks execute a traversals only yielding blocks loaded along the way
+export async function* walkBlocks(
   node: Node,
   sel: Selector,
   source: LinkLoader
@@ -786,7 +755,7 @@ export async function* walk(
       }
       const sNext = sel.explore(nd.value, ps);
       if (sNext !== null) {
-        yield* walk(value, sNext, source);
+        yield* walkBlocks(value, sNext, source);
       }
     }
   } else {
@@ -798,125 +767,8 @@ export async function* walk(
       }
       const sNext = sel.explore(nd.value, pathSegment);
       if (sNext !== null) {
-        yield* walk(new Node(value), sNext, source);
+        yield* walkBlocks(new Node(value), sNext, source);
       }
     }
   }
-}
-
-function parsePath(path: string): {root: CID; segments: string[]} {
-  const comps = toPathComponents(path);
-  const root = CID.parse(comps[0]);
-  return {
-    segments: comps,
-    root,
-  };
-}
-
-interface loaderFactory {
-  newLoader(root: CID, link: CID, sel: SelectorNode): LinkLoader;
-}
-
-export function offlineLoader(blocks: Blockstore) {
-  return {
-    newLoader(root: CID, link: CID, sel: SelectorNode) {
-      return {
-        load: (cid: CID) => blockFromStore(cid, blocks),
-        close: () => {},
-      };
-    },
-  };
-}
-
-/**
- * resolve content from a DAG using a path. May execute multiple data transfers to obtain the required blocks.
- */
-export async function* resolve(
-  path: string,
-  lf: loaderFactory
-): AsyncIterable<any> {
-  const {segments, root} = parsePath(path);
-  let cid = root;
-  let segs = segments.slice(1);
-  let isLast = false;
-
-  do {
-    if (segs.length === 0) {
-      isLast = true;
-    }
-    // for unixfs unless we know the index of the path we're looking for
-    // we must recursively request the entries to find the link hash
-    // a trailing slash at the end of a path will treat it as a directory
-    const sel = getSelector(
-      segs.length === 0 && path[path.length - 1] !== '/' ? '*' : '/'
-    );
-    const loader = lf.newLoader(root, cid, sel);
-    incomingBlocks: for await (const blk of walk(
-      new Node(cid),
-      parseContext().parseSelector(sel),
-      loader
-    )) {
-      // if not cbor or dagpb just return the bytes
-      switch (blk.cid.code) {
-        case 0x70:
-        case 0x71:
-          break;
-        default:
-          yield blk.bytes;
-          continue incomingBlocks;
-      }
-      try {
-        const unixfs = UnixFS.unmarshal(blk.value.Data);
-        if (unixfs.isDirectory()) {
-          // if it's a directory and we have a segment to resolve, identify the link
-          if (segs.length > 0) {
-            for (const link of blk.value.Links) {
-              if (link.Name === segs[0]) {
-                cid = link.Hash;
-                segs = segs.slice(1);
-                break incomingBlocks;
-              }
-            }
-            throw new Error('key not found: ' + segs[0]);
-          } else {
-            // if the block is a directory and we have no key return the entries as JSON
-            yield dagJSON.encode(
-              blk.value.Links.map((l: PBLink) => ({
-                name: l.Name,
-                hash: l.Hash.toString(),
-                size: l.Tsize,
-              }))
-            );
-            break incomingBlocks;
-          }
-        }
-        if (unixfs.type === 'file') {
-          if (unixfs.data && unixfs.data.length) {
-            yield unixfs.data;
-          }
-          continue incomingBlocks;
-        }
-      } catch (e) {}
-      // we're outside of unixfs territory
-      if (segs.length > 0) {
-        // best effort to access the field associated with the key
-        const key = segs[0];
-        const field = blk.value[key];
-        if (field) {
-          const link = CID.asCID(field);
-          if (link) {
-            cid = link;
-            segs = segs.slice(1);
-          } else {
-            yield field;
-          }
-        }
-      } else {
-        yield blk.bytes;
-        continue incomingBlocks;
-      }
-    }
-    // tell the loader we're done receiving blocks for this traversal
-    loader.close();
-  } while (!isLast);
 }
